requirements.txt
huggingface_hub
python-dotenv
tqdm
aiohttp
modelscope


